fix(admin): non-superuser access to classes
Previously, non-superusers have access to all classes instead of those
that belong to their headquater only. For example, an admin of bnu would
be able to see classes exclusive to bnuzh also, which is not troublesome
really but still quite annoying somewhat.

This fix deals with that, and now non-superusers have only access to
classes of their corresponding headquater.